feat(day06): implement part two

+++ src/main/java/org/example/day06/Race.java
@@ -1,4 +1,4 @@
 package org.example.day06;
 
-public record Race(int time, int distance) {
+public record Race(int time, long distance) {
 }
